# server/openapi.yaml
openapi: 3.1.0
info:
  title: Image Annotation Hybrid Project API
  version: 1.0.0
  description: |
    Remote API for hybrid image-annotation projects. Mutating requests are
    authenticated, authorized per project, audited, and exposed through the
    incremental change stream.
servers:
  - url: http://127.0.0.1:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/projects:
    get:
      tags: [Projects]
      operationId: listProjects
      responses:
        "200":
          description: Projects visible to the current user
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Project"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [Projects]
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectWrite"
      responses:
        "200":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [Projects]
      operationId: getProject
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        default:
          $ref: "#/components/responses/Error"
    patch:
      tags: [Projects]
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectWrite"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        default:
          $ref: "#/components/responses/Error"
    delete:
      tags: [Projects]
      operationId: deleteProject
      responses:
        "204":
          description: Project soft-deleted
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Structured API error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    ProjectRole:
      type: string
      enum: [owner, manager, annotator, reviewer, viewer]
    Project:
      type: object
      required: [id, name, mode, revision]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: [string, "null"]
        mode:
          type: string
          enum: [cloud_linked, mirrored]
        revision:
          type: integer
          format: int64
        role:
          $ref: "#/components/schemas/ProjectRole"
    ProjectWrite:
      type: object
      required: [name, mode]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: [string, "null"]
        mode:
          type: string
          enum: [cloud_linked, mirrored]
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: [string, "null"]
        details:
          type: [object, "null"]
          additionalProperties: true